requirements.txt
telethon
python-dotenv
psycopg2-binary
dbt-postgres
ultralytics
uvicorn